+++ lib/src/badge_position.dart
@@ -1,24 +1,60 @@
+import 'package:badges/badges.dart';
+
+/// Utility class. Don't try to use it!
+///
+/// This contain a custom position for the [Badge]
 class BadgePosition {
-  final double top;
-  final double right;
-  final double bottom;
-  final double left;
+  /// This is the distance to the top side of a parent widget.
+  final double? top;
 
-  const BadgePosition({this.top, this.right, this.bottom, this.left});
+  /// This is the distance to the end side of a parent widget.
+  final double? end;
 
-  factory BadgePosition.topLeft({double top, double left}) {
-    return BadgePosition(top: top ?? -5, left: left ?? -10);
+  /// This is the distance to the start side of a parent widget.
+  final double? start;
+
+  /// This is the distance to the bottom side of a parent widget.
+  final double? bottom;
+
+  /// This is indicate that centralise is required
+  final bool isCenter;
+
+  /// This creates a new instance of this widget
+  const BadgePosition({
+    this.top,
+    this.end,
+    this.bottom,
+    this.start,
+    this.isCenter = false,
+  });
+
+  /// Factory method that creates a new instance of this widget
+  /// according to the center
+  factory BadgePosition.center() {
+    return BadgePosition(isCenter: true);
   }
 
-  factory BadgePosition.topRight({double top, double right}) {
-    return BadgePosition(top: top ?? -8, right: right ?? -10);
+  /// Factory method that creates a new instance of this widget
+  /// according to [top] and [start]
+  factory BadgePosition.topStart({double top = -5, double start = -10}) {
+    return BadgePosition(top: top, start: start);
   }
 
-  factory BadgePosition.bottomRight({double bottom, double right}) {
-    return BadgePosition(bottom: bottom ?? -8, right: right ?? -10);
+  /// Factory method that creates a new instance of this widget
+  /// according to [top] and [end]
+  factory BadgePosition.topEnd({double top = -8, double end = -10}) {
+    return BadgePosition(top: top, end: end);
   }
 
-  factory BadgePosition.bottomLeft({double bottom, double left}) {
-    return BadgePosition(bottom: bottom ?? -8, left: left ?? -10);
+  /// Factory method that creates a new instance of this widget
+  /// according to [bottom] and [end]
+  factory BadgePosition.bottomEnd({double bottom = -8, double end = -10}) {
+    return BadgePosition(bottom: bottom, end: end);
+  }
+
+  /// Factory method that creates a new instance of this widget
+  /// according to [bottom] and [start]
+  factory BadgePosition.bottomStart({double bottom = -8, double start = -10}) {
+    return BadgePosition(bottom: bottom, start: start);
   }
 }