+++ lib/src/badge_shape.dart
@@ -1,1 +1,17 @@
-enum BadgeShape { circle, square }
+import 'package:badges/badges.dart';
+import 'package:flutter/material.dart';
+
+/// This is a set of shapes that you can use for your [Badge] widget.
+enum BadgeShape {
+  /// This value is for using a circular badges
+  ///
+  /// See also:
+  /// * [CircleBorder]
+  circle,
+
+  /// This value is for using a square badges
+  ///
+  /// See also:
+  /// * [RoundedRectangleBorder]
+  square,
+}